requirements.txt
asgiref==3.11.0
attrs==25.4.0
colorama==0.4.6
coverage==7.13.0
Django==5.2
django-environ==0.12.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pluggy==1.6.0
psycopg==3.3.2
psycopg-binary==3.3.2
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.4
tzdata==2025.2
uritemplate==4.2.0
